Create raspberry_pesticide_flow
This project is a smart agriculture system which works on deep learning and Raspberry Pi. It captures real-time images of plant leaves using a USB camera and processes them with a pre-trained CNN model to classify the leaf as healthy or infected. Based on the prediction, the Raspberry Pi controls a relay-driven pump that sprays pesticide automatically for a set duration. The system combines computer vision, IoT hardware (camera, relay, pump), and machine learning to reduce manual intervention, optimize pesticide usage, and support sustainable farming practices.